ご確認よろしくお願いします。

+++ src/main/java/jp/co/seattle/library/controller/DeleteBookController.java
@@ -13,6 +13,7 @@ import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RequestParam;
 
 import jp.co.seattle.library.service.BooksService;
+import jp.co.seattle.library.service.RentalsService;
 
 /**
  * 削除コントローラー
@@ -23,6 +24,9 @@ public class DeleteBookController {
 
 	@Autowired
 	private BooksService booksService;
+
+	@Autowired
+	private RentalsService rentalsService;
 
 	/**
 	 * 対象書籍を削除する
@@ -36,10 +40,18 @@ public class DeleteBookController {
 	@RequestMapping(value = "/deleteBook", method = RequestMethod.POST)
 	public String deleteBook(Locale locale, @RequestParam("bookId") Integer bookId, Model model) {
 		logger.info("Welcome delete! The client locale is {}.", locale);
-		booksService.deleteBook(bookId);
-		// 本の情報を取得して画面側に渡す
-		model.addAttribute("bookList", booksService.getBookList());
-		return "home";
+		// 対象書籍が貸出されていない時
+		if (rentalsService.getRentInfo(bookId) == 0) {
+			booksService.deleteBook(bookId);
+			// 本の情報を取得して画面側に渡す
+			model.addAttribute("bookList", booksService.getBookList());
+			return "home";
+			// 対象書籍が貸出されている時
+		} else {
+			model.addAttribute("bookDetailsInfo", booksService.getBookInfo(bookId));
+			model.addAttribute("errorMessage_rent_return", "貸出し中のため書籍を削除できません。");
+			return "details";
+		}
 
 	}
